法律法规追踪报告系统-gemini/requirements.txt
# 智能法律法规追踪系统 - 企业级依赖清单

# 现代化GUI框架
customtkinter==5.2.2

# 异步HTTP客户端
aiohttp==3.9.5
httpx==0.27.0

# 高性能HTML解析
selectolax==0.3.21

# AI服务
zhipuai>=2.1.0

# 数据库与缓存
redis==5.0.8

# 任务调度
apscheduler==3.10.4

# 日志系统
loguru==0.7.2

# 配置管理
pydantic==2.8.2

# 终端输出美化
rich==13.7.1

# 数据处理
pandas==2.2.2

# 数据可视化
plotly==5.22.0

# 环境变量管理
python-dotenv==1.0.1

# 重试机制
tenacity==8.5.0

# SQLite ORM
sqlalchemy==2.0.23

# 数据验证
marshmallow==3.21.0

# 邮件发送


# JSON处理
orjson==3.9.10

# 时间处理
pendulum==3.0.0

# 加密
cryptography==41.0.7

# 进度条
tqdm==4.66.1

# 开发工具
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1